+++ src/main/java/com/example/shinhanserver/domain/PB/PBRepository.java
@@ -1,5 +1,6 @@
 package com.example.shinhanserver.domain.PB;
 
+import com.example.shinhanserver.domain.entity.PB;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
