+++ src/main/java/com/example/shinhanserver/domain/entity/Career.java
@@ -1,37 +1,39 @@
-package com.example.shinhanserver.domain;
+package com.example.shinhanserver.domain.entity;
 
 import javax.persistence.*;
 
-import com.example.shinhanserver.domain.PB.PB;
+import com.example.shinhanserver.domain.entity.PB;
 import lombok.AccessLevel;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-
 import lombok.*;
 
-
 import java.time.LocalDate;
+
 
 @Getter
 @Entity
 @Builder
 @AllArgsConstructor
 @NoArgsConstructor(access= AccessLevel.PROTECTED)
-@Table(name = "certification")
-public class Certification {
+@Table(name = "career")
+public class Career {
 
   @Id
   @GeneratedValue(strategy = GenerationType.IDENTITY)
-  @Column(name = "certification_id")
+  @Column(name = "career_id")
   private Long id;
 
   @ManyToOne(fetch = FetchType.LAZY)
   @JoinColumn(name = "pb_id")
   private PB pb;
 
-  private String name;
+  private String organization;
 
-  private String publisher;
+  private String location;
 
-  private LocalDate date;
+  private LocalDate start_date;
+
+  private LocalDate end_date;
+
 }
