+++ src/main/java/com/example/shinhanserver/domain/entity/Certification.java
@@ -1,13 +1,14 @@
-package com.example.shinhanserver.domain;
+package com.example.shinhanserver.domain.entity;
 
 import javax.persistence.*;
 
-import com.example.shinhanserver.domain.PB.PB;
+import com.example.shinhanserver.domain.entity.PB;
 import lombok.AccessLevel;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
 import lombok.*;
+
 
 import java.time.LocalDate;
 
@@ -16,23 +17,21 @@ import java.time.LocalDate;
 @Builder
 @AllArgsConstructor
 @NoArgsConstructor(access= AccessLevel.PROTECTED)
-@Table(name = "education")
-public class Education {
+@Table(name = "certification")
+public class Certification {
 
   @Id
   @GeneratedValue(strategy = GenerationType.IDENTITY)
-  @Column(name = "education_id")
+  @Column(name = "certification_id")
   private Long id;
-
-  private String edu;
-
-  private String major;
-
-  private LocalDate start_date;
-
-  private LocalDate end_date;
 
   @ManyToOne(fetch = FetchType.LAZY)
   @JoinColumn(name = "pb_id")
   private PB pb;
+
+  private String name;
+
+  private String publisher;
+
+  private LocalDate date;
 }
