+++ src/main/java/com/example/shinhanserver/domain/entity/Education.java
@@ -1,39 +1,38 @@
-package com.example.shinhanserver.domain;
+package com.example.shinhanserver.domain.entity;
 
 import javax.persistence.*;
 
-import com.example.shinhanserver.domain.PB.PB;
+import com.example.shinhanserver.domain.entity.PB;
 import lombok.AccessLevel;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
+
 import lombok.*;
 
 import java.time.LocalDate;
-
 
 @Getter
 @Entity
 @Builder
 @AllArgsConstructor
 @NoArgsConstructor(access= AccessLevel.PROTECTED)
-@Table(name = "career")
-public class Career {
+@Table(name = "education")
+public class Education {
 
   @Id
   @GeneratedValue(strategy = GenerationType.IDENTITY)
-  @Column(name = "career_id")
+  @Column(name = "education_id")
   private Long id;
 
-  @ManyToOne(fetch = FetchType.LAZY)
-  @JoinColumn(name = "pb_id")
-  private PB pb;
+  private String edu;
 
-  private String organization;
-
-  private String location;
+  private String major;
 
   private LocalDate start_date;
 
   private LocalDate end_date;
 
+  @ManyToOne(fetch = FetchType.LAZY)
+  @JoinColumn(name = "pb_id")
+  private PB pb;
 }
