+++ src/main/java/com/example/shinhanserver/domain/entity/Specialization.java
@@ -1,4 +1,4 @@
-package com.example.shinhanserver.domain;
+package com.example.shinhanserver.domain.entity;
 
 
 import lombok.*;
