+++ src/main/java/com/example/shinhanserver/domain/entity/Transaction.java
@@ -1,4 +1,4 @@
-package com.example.shinhanserver.domain;
+package com.example.shinhanserver.domain.entity;
 
 import javax.persistence.*;
 import lombok.AccessLevel;
