+++ src/main/java/com/example/shinhanserver/domain/form/CareerRepository.java
@@ -1,6 +1,6 @@
-package com.example.shinhanserver.repository;
+package com.example.shinhanserver.domain.form;
 
-import com.example.shinhanserver.domain.Career;
+import com.example.shinhanserver.domain.entity.Career;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface CareerRepository extends JpaRepository<Career, Long> {
