+++ src/main/java/com/example/shinhanserver/domain/form/CertificationRepository.java
@@ -1,6 +1,6 @@
-package com.example.shinhanserver.repository;
+package com.example.shinhanserver.domain.form;
 
-import com.example.shinhanserver.domain.Certification;
+import com.example.shinhanserver.domain.entity.Certification;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface CertificationRepository extends JpaRepository<Certification, Long> {
