+++ src/main/java/com/example/shinhanserver/domain/form/EducationRepository.java
@@ -1,6 +1,6 @@
-package com.example.shinhanserver.repository;
+package com.example.shinhanserver.domain.form;
 
-import com.example.shinhanserver.domain.Education;
+import com.example.shinhanserver.domain.entity.Education;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface EducationRepository extends JpaRepository<Education, Long> {
