+++ src/main/java/com/example/shinhanserver/domain/pbinfo/PBPageService.java
@@ -1,7 +1,7 @@
 //package com.example.shinhanserver.service;
 //
 //import com.example.shinhanserver.domain.PB;
-//import com.example.shinhanserver.dto.PBPageDto;
+//import com.example.shinhanserver.domain.pbinfo.PBPageDto;
 //import com.example.shinhanserver.repository.PBRepository;
 //import lombok.AllArgsConstructor;
 //import lombok.extern.slf4j.Slf4j;
